avoid overflow in mapped test

+++ util_benchsrLarge.cpp
@@ -127,8 +127,8 @@ void sendRecvMapped(MPI_Comm mpiComm, const bool isRdv, const int mbpr,
   std::stringstream ss;
   ss << mbpr << " B " << name;
   if(!isRdv) { //sender
-    adios2::Dims shape{static_cast<size_t>(mbpr*nproc)};
-    adios2::Dims start{static_cast<size_t>(mbpr*rank)};
+    adios2::Dims shape{static_cast<size_t>(mbpr)*nproc};
+    adios2::Dims start{static_cast<size_t>(mbpr)*rank};
     adios2::Dims count{static_cast<size_t>(mbpr)};
     auto var = io.DefineVariable<redev::LO>(name, shape, start, count);
     assert(var);
@@ -151,7 +151,7 @@ void sendRecvMapped(MPI_Comm mpiComm, const bool isRdv, const int mbpr,
     auto msgs = io.InquireVariable<redev::LO>(name);
     assert(msgs);
     redev::LOs inMsgs(mbpr);
-    const auto startRead = mbpr*reductionFactor*rank;
+    const auto startRead = static_cast<size_t>(mbpr)*reductionFactor*rank;
     msgs.SetSelection({{static_cast<size_t>(startRead)},
                        {static_cast<size_t>(mbpr)}
                       });
